+++ packages/assert/src/__tests__/object.assert.test.ts
@@ -1,4 +1,4 @@
-import { errorMessages } from '../../messages/errorMessages';
+import { errorMessages } from '../messages/errorMessages';
 import { assertPlainObject } from '../object.asserts';
 
 describe('object assertions tests', () => {
