+++ packages/assert/src/barcode.types.ts
