+++ packages/assert/src/string.guards.ts
