--- packages/assert/src/types/index.ts
@@ -1,3 +1,0 @@
-export type * from './barcode.types';
-export type * from './object.types';
-export type * from './uuid.types';
